checkpoint: moves "*/StaticConfigReadingError.ts" to "*/Reading/Error.ts"
