+++ src/features/TextToImage/Config/Static/Reading/exports.ts
@@ -1,3 +1,3 @@
 export {
   default as Error,
-} from '../StaticConfigReadingError';
+} from './Error';
